feat: Add tests for asserting compute budget limits (#23)
# Description

Add common test setup for each transaction type the client sends. Add tests for each transaction type which utilize `arbtests` to assert that the budgets are high enough.
Adjust the `justfile` to contain these tests as a separate target.

## Summary

Add tests for each transaction we are creating, ensuring we are setting the correct CU budget limits.

## Testing

This PR adds tests for asserting correctness of the `COMPUTE_BUDGET_LIMIT` constants.

# P.S.

Sorry about the large diff - most of it is just the lockfile