feat: Explicitly pass in 0 retries (#63)
# Description

Sets `max_retries` to `Some(0)` in the `RpcSendTransactionConfig` when sending transactions, effectively disabling automatic retries. Previously, this was set to `None` which would use the default retry behavior.
Based on [Helius guides](https://docs.helius.dev/guides/sending-transactions-on-solana) this is the better option for having a response from the network sooner.