fix: Update CLI build matrix and use dynamic runner OS (#79)
# Description

This PR updates the GitHub workflow for building the CLI. It removes the `x86_64-apple-darwin` target from the build matrix while keeping the `aarch64-apple-darwin` and `x86_64-unknown-linux-gnu` targets. Additionally, it moves the `runs-on` configuration from a fixed value to a dynamic one that uses the OS specified in each matrix entry.