fix: Manually instantiate rustls CryptoProvider due to aws-sdk-s3 conflict (#213)
# Description

Added AWS-LC as the cryptographic provider for rustls in the Yellowstone consumer. This change ensures proper TLS functionality by explicitly setting the cryptographic backend for rustls to AWS-LC.

Also updated the indexer CLI arguments by changing the short flag for logpath and updating the justfile to include the required S3 bucket parameter when running the indexer.

- [x] Are all new dependencies necessary in `Cargo.toml` necessary