chore: Benchmark cleanup (#312)
# Description

This PR enhances the benchmark functionality by properly initializing and closing the blober before and after benchmark runs. It also reduces the compute unit limit for the `FinalizeBlob` transaction from 25,000 to 5,000 units to optimize resource usage.

Additionally, the PR simplifies the `run-benchmark` command in the justfile by removing the indexer URL parameter and using a more straightforward command structure with a default data path.