chore: Deprecate old proofs methods (#359)
# Description

This PR deprecates the `get_proof` and `get_proof_for_blob` methods in the indexer API, as well as the corresponding CLI commands. Users are now directed to use the `zk-proof` command and `checkpoint_proof` method instead.

The changes include:
- Adding `#[deprecated]` attributes to the `ProofForBlob` and `Proof` CLI subcommands
- Adding `#[deprecated]` attributes to the `get_proof` and `get_proof_for_blob` methods in the client
- Adding `#[deprecated]` attributes to the corresponding RPC methods
- Adding `#[allow(deprecated)]` annotations where needed to suppress compiler warnings

## Linked Issues
- Related to the transition to the new proof system

## Testing
The changes are non-functional and only add deprecation notices. Existing functionality continues to work as before.